feat: Se completa CRUD de roles con notificaciones

+++ InversionesZJ/InversionesZJ.Domain/Interfaces/Common/IGenericRepository.cs
@@ -7,7 +7,7 @@ namespace InversionesZJ.Domain.Interfaces.common;
 
 public interface IGenericRepository<T> where T : BaseEntity
 {
-    Task<T?> GetByIdAsync(int id, CancellationToken ct = default);
+    Task<T?> GetByIdAsync(long  id, CancellationToken ct = default);
     Task<List<T>> GetAllAsync(CancellationToken ct = default);
     Task AddAsync (T entity, CancellationToken ct = default);
     Task UpdateAsync (T entity, CancellationToken ct = default);
